Convert GoLink into builder action
